NLP-llm-server/requirements.txt
fastapi
pydantic
ollama
uvicorn
sentence-transformers
sentence-transformers
scikit-learn
groq
python-dotenv
